+++ CodexLib/Components/ActivatableVariants.cs
@@ -1,5 +1,4 @@
 ﻿using Kingmaker.UI;
-using Kingmaker.UI.MVVM._VM.ActionBar;
 using Kingmaker.UI.UnitSettings;
 using System;
 using System.Collections.Generic;
@@ -9,50 +8,38 @@ using System.Threading.Tasks;
 
 namespace CodexLib
 {
-    internal class ActivatableVariants : UnitFactComponentDelegate<ActivatableVariantsData>, IActionBarConvert
+    public class ActivatableVariants : UnitFactComponentDelegate<VariantSelectionData>, IActionBarConvert
     {
-        public BlueprintUnitFact[] Facts;
+        public bool NeedFact = true;
+        public BlueprintUnitFactReference[] Facts;
 
-        public List<MechanicActionBarSlot> GetConverts() // UnitFact? BlueprintUnitFact?
+        public ActivatableVariants(params AnyRef[] facts)
         {
-            //return this.Facts;
-            throw new NotImplementedException();
+            this.Facts = facts.ToRef<BlueprintUnitFactReference>();
+        }
+
+        public List<MechanicActionBarSlot> GetConverts()
+        {
+            var data = this.Data;
+
+            var list = new List<MechanicActionBarSlot>();
+            foreach (var fact in this.Facts)
+            {
+                if (fact.NotEmpty() && (!this.NeedFact || this.Owner.HasFact(fact)))
+                    list.Add(new MechanicActionBarSlotVariantSelection(this.Owner, fact, data));
+            }
+
+            return list;
+        }
+
+        public Sprite GetIcon()
+        {
+            return this.Data.Selected?.Get()?.Icon;
         }
     }
 
-    internal class ActivatableVariantsData
+    public class VariantSelectionData
     {
-
-    }
-
-    [HarmonyPatch]
-    internal class Patch_ActionBarConvert
-    {
-        [HarmonyPatch(typeof(ActionBarSlotVM), nameof(ActionBarSlotVM.SetMechanicSlot))]
-        [HarmonyPostfix]
-        public static void SetMechanicSlot(ActionBarSlotVM __instance)
-        {
-            if (__instance.MechanicActionBarSlot.GetContentData() is UnitFact fact && fact.Blueprint.GetComponent<IActionBarConvert>() != null)
-                __instance.HasConvert.Value = true;
-        }
-
-        [HarmonyPatch(typeof(ActionBarSlotVM), nameof(ActionBarSlotVM.OnShowConvertRequest))]
-        [HarmonyPriority(Priority.VeryHigh)]
-        [HarmonyPrefix]
-        public static bool OnShowConvertRequest(ActionBarSlotVM __instance)
-        {
-            if (__instance.ConvertedVm.Value != null && !__instance.ConvertedVm.Value.IsDisposed)
-                return true;
-
-            if (__instance.MechanicActionBarSlot.GetContentData() is not UnitFact fact)
-                return true;
-
-            var convert = fact.Blueprint.GetComponent<IActionBarConvert>();
-            if (convert == null)
-                return true;
-
-            __instance.ConvertedVm.Value = new ActionBarConvertedVMAny(__instance, convert.GetConverts(), __instance.CloseConvert);
-            return false;
-        }
+        public BlueprintUnitFactReference Selected;
     }
 }