+++ DarkCodex/Properties/AssemblyInfo.cs
@@ -31,6 +31,6 @@ using System.Runtime.InteropServices;
 //
 // Sie können alle Werte angeben oder Standardwerte für die Build- und Revisionsnummern verwenden,
 // indem Sie "*" wie unten gezeigt eingeben:
-// [assembly: AssemblyVersion("1.3.20")]
-[assembly: AssemblyVersion("1.3.20")]
-[assembly: AssemblyFileVersion("1.3.20")]
+// [assembly: AssemblyVersion("1.3.21")]
+[assembly: AssemblyVersion("1.3.21")]
+[assembly: AssemblyFileVersion("1.3.21")]